requirements.txt
numpy>=1.23.4
pandas>=1.5.2
matplotlib>=3.6.2
opencv-python>=4.6.0.66
astropy>=5.2.2
Pillow>=9.2.0
tqdm>=4.64.0
tensorflow>=2.7.1
pyyaml>=6.0
torch>=1.13.0
scikit-learn>=1.2.2
torchvision>=0.14.0
